+++ packages/experimental/Checkbox/src/shouldUseThickCheckmark.ts
@@ -1,0 +1,5 @@
+import { Theme } from '@fluentui-react-native/framework';
+
+export function shouldUseThickCheckmark(_theme: Theme): boolean {
+  return false;
+}