+++ packages/components/FocusZone/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/focus-zone
 
-This log was last generated on Fri, 11 Mar 2022 00:34:53 GMT and should not be manually modified.
+This log was last generated on Mon, 21 Mar 2022 01:47:54 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.10
+
+Mon, 21 Mar 2022 01:47:54 GMT
+
+### Patches
+
+- Fix typo (sanajmi@microsoft.com)
 
 ## 0.9.9